+++ sdk/attestation/azure-security-attestation/inc/azure/attestation/attestation_client_options.hpp
@@ -16,28 +16,6 @@
 
 namespace Azure { namespace Security { namespace Attestation {
 
-  /** @brief Version to be used when communicating with the Attestation service.
-   */
-  class ServiceVersion final
-      : public Azure::Core::_internal::ExtendableEnumeration<ServiceVersion> {
-  public:
-    /**
-     * @brief Construct a new Service Version object
-     *
-     * @param version The string version for the Attestation service.
-     */
-    explicit ServiceVersion(std::string version)
-        : Azure::Core::_internal::ExtendableEnumeration<ServiceVersion>(std::move(version))
-    {
-    }
-
-    /**
-     * @brief Use to send request to the 2020-10-01 version of Attestation service.
-     *
-     */
-    AZ_ATTESTATION_DLLEXPORT static const ServiceVersion V2020_10_01;
-  };
-
   /**
    * @brief The TokenValidationCallbackFn represents a callback which is called to allow the caller
    *  to perform additional token validation options beyond the validations performed by the
@@ -116,7 +94,7 @@ namespace Azure { namespace Security { namespace Attestation {
   {
     /** @brief Version to use when communicating with the attestation service.
      */
-    ServiceVersion Version;
+    std::string ApiVersion{"2020-10-01"};
 
     /** @brief Options sent when validating tokens received by the attestation service.
      */
@@ -125,15 +103,11 @@ namespace Azure { namespace Security { namespace Attestation {
     /**
      * @brief Construct a new Attestation Client Options object.
      *
-     * @param version Optional version for the client.
      * @param tokenValidationOptions Options applied when validating attestation tokens returned by
      * the service.
      */
-    AttestationClientOptions(
-        ServiceVersion version = ServiceVersion::V2020_10_01,
-        AttestationTokenValidationOptions const& tokenValidationOptions = {})
-        : Azure::Core::_internal::ClientOptions(), Version(version),
-          TokenValidationOptions(tokenValidationOptions)
+    AttestationClientOptions(AttestationTokenValidationOptions const& tokenValidationOptions = {})
+        : Azure::Core::_internal::ClientOptions(), TokenValidationOptions(tokenValidationOptions)
     {
     }
   };
@@ -145,22 +119,19 @@ namespace Azure { namespace Security { namespace Attestation {
   {
     /** @brief Version to use when communicating with the attestation service.
      */
-    ServiceVersion Version;
+    std::string ApiVersion{"2020-10-01"};
     /** @brief Options sent when validating tokens received by the attestation service.
      */
     AttestationTokenValidationOptions TokenValidationOptions;
     /**
      * @brief Construct a new Attestation Client Options object.
      *
-     * @param version Optional version for the client.
      * @param tokenValidationOptions Options applied when validating attestation tokens returned by
      * the service.
      */
     AttestationAdministrationClientOptions(
-        ServiceVersion version = ServiceVersion::V2020_10_01,
         AttestationTokenValidationOptions const& tokenValidationOptions = {})
-        : Azure::Core::_internal::ClientOptions(), Version(version),
-          TokenValidationOptions(tokenValidationOptions)
+        : Azure::Core::_internal::ClientOptions(), TokenValidationOptions(tokenValidationOptions)
     {
     }
   };