+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/cryptography_client.cpp
@@ -97,7 +97,7 @@ CryptographyClient::CryptographyClient(
     std::string const& keyId,
     std::shared_ptr<Core::Credentials::TokenCredential const> credential,
     CryptographyClientOptions const& options)
-    : m_keyId(Azure::Core::Url(keyId)), m_apiVersion(options.Version.ToString())
+    : m_keyId(Azure::Core::Url(keyId)), m_apiVersion(options.Version)
 {
   std::vector<std::unique_ptr<HttpPolicy>> perRetrypolicies;
   {