+++ sdk/keyvault/azure-security-keyvault-secrets/test/ut/secret_client_test.cpp
@@ -28,20 +28,10 @@ TEST(SecretClient, ServiceVersion)
 {
   auto credential
       = std::make_shared<Azure::Identity::ClientSecretCredential>("tenantID", "AppId", "SecretId");
-  {
-    // 7.2
-    EXPECT_NO_THROW(
-        auto options = SecretClientOptions(ServiceVersion::V7_2);
-        SecretClient SecretClient("http://account.vault.azure.net", credential, options);
-        EXPECT_EQ(options.Version.ToString(), "7.2"););
-  }
-  {
-    // arbitrary version
-    EXPECT_NO_THROW(
-        auto options = SecretClientOptions(ServiceVersion("1.0"));
-        SecretClient secretClient("http://account.vault.azure.net", credential, options);
-        EXPECT_EQ(options.Version.ToString(), "1.0"););
-  }
+  // 7.3
+  EXPECT_NO_THROW(auto options = SecretClientOptions();
+                  SecretClient SecretClient("http://account.vault.azure.net", credential, options);
+                  EXPECT_EQ(options.ApiVersion, "7.3"););
 }
 
 TEST(SecretClient, GetUrl)