+++ sdk/attestation/azure-security-attestation/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.4 (Unreleased)
+## 1.1.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,12 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0 (2022-07-07)
+
+### Breaking Changes
+
+- Renamed `Version` field to `ApiVersion` and removed the `ServiceVersion` enumeration.
 
 ## 1.0.0-beta.3 (2022-06-07)