+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -1,16 +1,20 @@
 # Release History
 
-## 1.4.0-beta.1 (Unreleased)
+## 1.4.0-beta.2 (Unreleased)
 
 ### Features Added
-
-- Added Azure App Service API version `2019-08-01` support for `ManagedIdentityCredential`.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.4.0-beta.1 (2022-06-30)
+
+### Features Added
+
+- Added Azure App Service API version `2019-08-01` support for `ManagedIdentityCredential`.
 
 ## 1.3.0 (2022-06-07)